[MOD] id Long 타입으로 변경

+++ spring/umc-spring-study/src/main/java/com/example/umc/repository/MemberMissionRepository.java
@@ -5,6 +5,6 @@ import com.example.umc.domain.Mission;
 import com.example.umc.domain.mapping.MemberMission;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MemberMissionRepository extends JpaRepository<MemberMission, Integer> {
+public interface MemberMissionRepository extends JpaRepository<MemberMission, Long> {
     boolean existsByMemberAndMission(Member member, Mission mission);
 }
